# feat: add efficient access to embeddings