--- a/.github/2.0/cookbooks/Document.md
+++ b/.github/2.0/cookbooks/Document.md
@@ -1042,7 +1042,6 @@ This can be very useful when extracting a batch of embeddings:
 
 ```python
 import numpy as np
-
 np.stack(da.get_attributes('embedding'))
 ```
 
@@ -1051,6 +1050,23 @@ np.stack(da.get_attributes('embedding'))
  [4 5 6]
  [7 8 9]]
 ```
+
+### Property `.embeddings`
+
+There is a faster version to extract embeddings from a DocumentArray or DocumentArrayMemmap, the property `.embeddings`. This property assumes all embeddings in the array have the same dimension. Note that
+
+```
+da.embeddings
+```
+
+will produce the same output as `np.stack(da.get_attributes('embedding'))` but the results will be retrieved faster.
+
+```
+[[1 2 3]
+ [4 5 6]
+ [7 8 9]]
+```
+
 
 
 ### Finding closest documents between `DocumentArray` objects