--- a/tests/unit/types/arrays/test_neural_ops.py
+++ b/tests/unit/types/arrays/test_neural_ops.py
@@ -395,27 +395,27 @@ def test_match_inclusive_dam(tmpdir):
     assert len(list(traversed)) == 9
 
 
-def test_da_embeddings(tmpdir):
-    da = DocumentArrayMemmap(tmpdir)
-    da.extend(Document(embedding=np.array([1, 2, 3, 4])) for _ in range(100))
+def test_da_get_embeddings():
+    da = DocumentArray(random_docs(100))
     np.testing.assert_almost_equal(da.get_attributes('embedding'), da.embeddings)
 
 
-def test_da_embeddings_slice(tmpdir):
-    da = DocumentArrayMemmap(tmpdir)
-    da.extend(Document(embedding=np.array([1, 2, 3, 4])) for _ in range(100))
+def test_dam_embeddings(tmpdir):
+    dam = DocumentArrayMemmap(tmpdir)
+    dam.extend(Document(embedding=np.array([1, 2, 3, 4])) for _ in range(100))
+    np.testing.assert_almost_equal(dam.get_attributes('embedding'), dam.embeddings)
+
+
+def test_da_get_embeddings():
+    da = DocumentArray(random_docs(100))
     np.testing.assert_almost_equal(
-        da.get_attributes('embedding')[10:20], da.embeddings[10:20]
+        da.get_attributes('embedding')[10:20], da._get_embeddings(slice(10, 20))
     )
 
 
-def test_da_embeddings():
-    da = DocumentArray(random_docs(100))
-    np.testing.assert_almost_equal(da.get_attributes('embedding'), da.embeddings)
-
-
-def test_da_embeddings_slice():
-    da = DocumentArray(random_docs(100))
+def test_dam_get_embeddings(tmpdir):
+    da = DocumentArrayMemmap(tmpdir)
+    da.extend(Document(embedding=np.array([1, 2, 3, 4])) for _ in range(100))
     np.testing.assert_almost_equal(
-        da.get_attributes('embedding')[10:20], da.embeddings[10:20]
+        da.get_attributes('embedding')[10:20], da._get_embeddings(slice(10, 20))
     )