This PR adds a `.embeddings` property for faster acces to the embeddings of a `DocumentArray` or a `DocumentArrayMemmap` . This is done iterating over the bytes of the embeddings at proto level avoiding the creation of intermediate Python obejcts. 

This PR also provides
- A speed up of 3x on the current `.match` implementation
- A speed up of 2x on the current `.match` implementation with minibatches

### Example:

```
from jina import Document, DocumentArray
import numpy as np
d = Document(embedding=np.random.random(128))
docarray = DocumentArray([d for _ in range(100_000)])
```

Using `get_attributes`
```
%timeit np.vstack(docarray.get_attributes('embedding'))
```
```
1.65 s ± 95.8 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

Using  `.embeddings`
```
%timeit docarray.embeddings
```
```
649 ms ± 16.8 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```


### Impact on `.match`

```
d1 = Document(embedding=np.array([0,0,0,0,1]))
da_1 = DocumentArray([d1])
da_2 = DocumentArray([Document(embedding=x) for x in np.random.random((1_000_000, 5))])
```

This PR
```
%time da_1.match(da_2, metric=metric, limit=3)
5.54 s ± 209 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```
Current master branch
```
%time da_1.match(da_2, metric=metric, limit=3)
15.2 s ± 400 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```


### Impact on `.match` with batching

This PR also provides `_get_embeddings` which can be used to get a slice of the embeddings which is relevant for computing in minibatches.

With the current master branch
```
dam = DocumentArrayMemmap('./my-memmap')
dam.extend([Document(embedding=x) for x in np.random.random((100_000, 5)) ])
%timeit da_1.match(dam, metric=metric, limit=3, batch_size=10_000)
```
```
5.03 s ± 278 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```
In this PR
```
%timeit da_1.match(dam, metric=metric, limit=3, batch_size=10_000)
```
```
3.26 s ± 159 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```


### Warning 

We have to give user responsability to construct Documents in the DocumentArray  with  embeddings of the same dtype and shape. 

We need to warn users to construct DocumentArrays with embeddings of the same type.
```
d1 = Document(embedding=np.array([1,2,3]))
d2 = Document(embedding=np.array([2.2,2.,5.])
aux = DocumentArray([d1,d2] )
aux.embeddings
array([[                  1,                   2,                   3],
       [4612136378390124954, 4611686018427387904, 4617315517961601024]])
```